moving mpi_service into database package
